+++ src/mod_auth.c
@@ -539,8 +539,8 @@ static const char *_authn_gettoken(const _mod_auth_ctx_t *ctx, http_message_t *r
 }
 
 int authn_checksignature(const char *key,
-		const char *data, int datalen,
-		const char *sign, int signlen)
+		const char *data, size_t datalen,
+		const char *sign, size_t signlen)
 {
 	if (hash_macsha256 != NULL && key != NULL)
 	{
@@ -575,7 +575,7 @@ int authn_checktoken(_mod_auth_ctx_t *ctx, const char *token)
 	const char *sign = strrchr(string, '.');
 	if (sign != NULL)
 	{
-		int datalen = sign - data;
+		size_t datalen = sign - data;
 		sign++;
 		if (authn_checksignature(mod->authn->config->secret, data, datalen, sign, strlen(sign)) == ESUCCESS)
 		{
